workers/requirements.txt
mlflow==2.8.1
kafka-python==2.0.2
psycopg2-binary==2.9.9
numpy==1.24.3
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
